// src/shared/api/models/getApiInsightsSalesType200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiInsightsSalesType200Data } from './getApiInsightsSalesType200Data';

export type GetApiInsightsSalesType200 = {
    data: GetApiInsightsSalesType200Data;
    message: string;
};