// src/shared/api/models/getApiInsightsSalesType200DataRecordsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiInsightsSalesType200DataRecordsItemProduct } from './getApiInsightsSalesType200DataRecordsItemProduct';
import type { GetApiInsightsSalesType200DataRecordsItemService } from './getApiInsightsSalesType200DataRecordsItemService';
import type { GetApiInsightsSalesType200DataRecordsItemBookingType } from './getApiInsightsSalesType200DataRecordsItemBookingType';
import type { GetApiInsightsSalesType200DataRecordsItemEmployee } from './getApiInsightsSalesType200DataRecordsItemEmployee';

export type GetApiInsightsSalesType200DataRecordsItem = {
    id: string;
    invoiceId: string;
    date: string;
    product: GetApiInsightsSalesType200DataRecordsItemProduct;
    service: GetApiInsightsSalesType200DataRecordsItemService;
    giftCards: number;
    punchCards: number;
    credited: number;
    giftCardUse: number;
    punchCardUse: number;
    paidAmount: number;
    bookingType?: GetApiInsightsSalesType200DataRecordsItemBookingType;
    employee: GetApiInsightsSalesType200DataRecordsItemEmployee;
};