// src/shared/api/models/getApiInsightsSalesTypeBookingType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiInsightsSalesTypeBookingType =
    (typeof GetApiInsightsSalesTypeBookingType)[keyof typeof GetApiInsightsSalesTypeBookingType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiInsightsSalesTypeBookingType = {
    all: 'all',
    manual: 'manual',
    online: 'online',
} as const;