// src/shared/api/models/getApiInsightsSalesTypeSortBy.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiInsightsSalesTypeSortBy =
    (typeof GetApiInsightsSalesTypeSortBy)[keyof typeof GetApiInsightsSalesTypeSortBy];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiInsightsSalesTypeSortBy = {
    date: 'date',
    employee: 'employee',
    amount: 'amount',
} as const;